[Microservices] organize coffee service lambda functions
