+++ MicroservicesRestAPIServerless/microservices/services/coffee/functions/getCoffeeList/index.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const { getDdbItemList } = require("../lib/helper/awsDDBHelper");
-const { extractUserDetails } = require("../lib/helper/auth");
-const { handleSuccess, handleError } = require("../lib/helper/responseHelper");
-const config = require("../lib/utils/config");
+const { getDdbItemList } = require("../../lib/helper/awsDDBHelper");
+const { extractUserDetails } = require("../../lib/helper/auth");
+const { handleSuccess, handleError } = require("../../lib/helper/responseHelper");
+const config = require("../../lib/utils/config");
 
 module.exports.handler = async (event) => {
   console.log("Get Coffee List : Start");
