+++ MicroservicesRestAPIServerless/microservices/services/coffee/functions/insertCoffee/index.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const { extractUserDetails } = require("../lib/helper/auth");
-const { insertDdbItem } = require("../lib/helper/awsDDBHelper");
-const { handleSuccess, handleError } = require("../lib/helper/responseHelper");
-const config = require("../lib/utils/config");
+const { extractUserDetails } = require("../../lib/helper/auth");
+const { insertDdbItem } = require("../../lib/helper/awsDDBHelper");
+const { handleSuccess, handleError } = require("../../lib/helper/responseHelper");
+const config = require("../../lib/utils/config");
 
 module.exports.handler = async (event) => {
   console.log("Insert Coffee : Start");
